Rename util file
