+++ src/gql/campaign-types.d.ts
@@ -1,4 +1,4 @@
-import { CampaignTypeCategory } from 'components/campaign-types/Categories'
+import { CampaignTypeCategory } from 'components/campaign-types/categories'
 import { UUID } from './types'
 
 export type CampaignTypesType = {
